Only post PR comments when jobs have finished
